+++ tilings/strategies/fusion/fusion.py
@@ -148,13 +148,7 @@ class FusionStrategy(Strategy[Tiling, GriddedPerm]):
         return False
 
     def is_two_way(self, comb_class: Tiling):
-        algo = self.fusion_algorithm(comb_class)
-        new_ass = algo.new_assumption()
-        fused_assumptions = (
-            ass.__class__(gps)
-            for ass, gps in zip(comb_class.assumptions, algo.assumptions_fuse_counters)
-        )
-        return new_ass in fused_assumptions
+        return False
 
     def constructor(
         self, comb_class: Tiling, children: Optional[Tuple[Tiling, ...]] = None